requirements.txt
python-dotenv~=1.0.0
llama-index==0.9.19
openai==1.6.0
transformers==4.36.2
accelerate==0.25.0
chromadb==0.4.21
google-generativeai==0.1.0rc2
cryptography==41.0.7
pypdf==3.17.3